+++ backend/src/main/java/com/autoride/dto/response/UserResponse.java
@@ -1,8 +1,10 @@
 package com.autoride.dto.response;
 
+import java.math.BigDecimal;
+
 public record UserResponse(
         Long id,
         String name,
         String email,
-        Double balance
+        BigDecimal balance
 ) {}
